+++ src/components/loadBtn.js
@@ -1,0 +1,4 @@
+const returnLoadBtn = () => {
+  return `<button class="load-more" type="button">load more</button>`;
+};
+export {returnLoadBtn};
